Hardcode Cargo target information
This explicitly sets the target information for both the zwreec library
and binary. It also disables the tests and documentation for (only) the
binary, thus preventing awkward and unintuitive warnings when running
tests. For example, Cargo itself uses the same approach with its binary.

Since the logging library was moved from `utils` to `bin` in commit
c403771, none of its tests will be run (and no documentation generated).
As far as I see, there are two reasonable alternatives: Move the logger
back into the core library and accept its presence there, or make it its
own crate. (I would prefer the latter)