Fix lexer tests. They were passing even though they were bad.
This is because 037ac8d implemented PartialEq for Token in a way that token
content was completely ignored. This is not suitable for testing. Therefore the
implementation was renamed to

pub is_same_token() -> bool

and any checks for equality were switched to this function. Then we could
derive PartialEq again to ensure proper behavior for testing. A regression test
was added to make sure this bug does not occur again.