Adds inheritance support to rustlex rules and simplify rustlex rules.
It works like this: Lines inside conditions starting with a colon (:) will be
treated as an include rule. If a rule with this name already exists its
contents will be inserted at the location of the colon-directive. This allows
better rule-reuse and prevents bugs that occur when changing a rule in one
place and forgetting to change them in another.