Add ALL missing documentation.
There can always be more doc tests and guidance and this issue is probably
never solved. But now everything at least has a short (or long) doc string
describing its use.

This enables the #[warn(missing_docs)] lint which generates a warning if any
module, function, struct, struct field, enum, enum variant or anything else is
not documented with a doc string. It is only ignored in very few selected
locations like the generated rustlex file and struct definitions in enums where
the struct fields are all the same and/or self-explanatory.

This updates rustlex to a version that disables missing_docs for the generated
lexer definition. You are welcome <3.

Fixes #212